Create ShowingManager

+++ src/main/java/ru/ilezzov/showregion/managers/regions/Region.java
@@ -2,6 +2,6 @@ package ru.ilezzov.showregion.managers.regions;
 
 import net.kyori.adventure.bossbar.BossBar;
 
-public record Region(String name, String actionBarText, String bossBarText, double bossBarProgress,
+public record Region(String name, String actionBarText, String bossBarText, float bossBarProgress,
                      BossBar.Color bossBarColor, BossBar.Overlay bossBarOverlay) {
 }
